Hovertank and 2.0
The second tank has a turret that is a separate object in the model. I
added materials to the various parts of the tank, including plating, gun
rims and the inside of the gun for firing effect if you want. Also, I
noticed a problem with the orientation of my tank, it is facing the
wrong way. I'll need to learn from you how to avoid that and fix it.
Hope you like it. I had to reduce the size of the tanks by half for them
to make sense in the game.