Changed the original Hovertank and fixed textures and 2.0
Added materials to the original hovertank and split the Hover icospheres
and boxes. Also put the 2.0 as an fbx and it's displaying the right
textures. Let me know what you think. I'll do a turret for the original
hovertank and then work on App Inventor to see if I can get my head
around it and code avengers as well, so I am ready for tomorrow.