New tank has materials arranged properly.
And UV mapping done so that the textures will apply to the correct parts
of the tank. Feel free to modify anything. But I can't replace the old
model with the new one. When I delete the old one, I loose the new one
as well, even after I placed the new one in a prefab.
Also, I really love your Hovertank. It's super cool. It was in Scene 1,
not scene 2, right? Or something got mixed up during the transfer.